+++ src/main/java/managers/OrderManager.java
@@ -1,8 +1,5 @@
-package r;
+package managers;
 
-import com.sun.corba.se.impl.resolver.ORBDefaultInitRefResolverImpl;
-import com.sun.org.apache.xpath.internal.operations.Or;
-import r.model.Order;
 import org.json.*;
 import java.io.*;
 import java.nio.file.Files;
@@ -146,5 +143,45 @@ public class OrderManager {
 
 
 
+    }
+
+    public static class Order {
+        private int orderId;
+        private int clientId;
+        private int staffId;
+        private String status;
+
+        public int getOrderId() {
+            return orderId;
+        }
+
+        public int getClientId() {
+            return clientId;
+        }
+
+        public int getStaffId() {
+            return staffId;
+        }
+
+        public String getStatus() {
+            return status;
+        }
+
+        public Order(int orderId, int clientId, int staffId, String status) {
+            this.orderId = orderId;
+            this.clientId = clientId;
+            this.staffId = staffId;
+            this.status = status;
+        }
+
+        @Override
+        public String toString() {
+            return "{" +
+                    " \"orderId\": " + orderId +
+                    ", \"clientId\": " + "\"" + clientId + "\"" +
+                    ", \"staffId\": " + "\"" + staffId + "\"" +
+                    ", \"status\": " + "\"" + status + "\""  +
+                    '}';
+        }
     }
 }
