+++ src/main/java/managers/ProductManager.java
@@ -1,6 +1,5 @@
-package r;
+package managers;
 
-import r.model.Product;
 import org.json.*;
 
 import java.io.*;
@@ -148,5 +147,60 @@ public class ProductManager {
             //System.out.println(manageStuff.searchJsonObj("Dyson Hair Curler"));
 
 
+    }
+
+    public static class Product {
+        private int productId;
+        private String name;
+        private int quantity;
+        private String description;
+        private int orderId;
+        private String category;
+
+        public String getCategory() {
+            return category;
+        }
+
+        public int getOrderId() {
+            return orderId;
+        }
+
+        public int getProductId() {
+            return productId;
+        }
+
+        public String getName() {
+            return name;
+        }
+
+        public int getQuantity() {
+            return quantity;
+        }
+
+        public String getDescription() {
+            return description;
+        }
+
+        public Product(int productId,int orderId,String category,String name, int quantity, String description) {
+            this.productId = productId;
+            this.orderId=orderId;
+            this.category=category;
+            this.name = name;
+            this.quantity = quantity;
+            this.description = description;
+
+        }
+
+        @Override
+        public String toString() {
+            return "{" +
+                    " \"productId\": " + productId +
+                    ", \"orderId\": "  + orderId  +
+                    ", \"category\": " + "\"" + category + "\"" +
+                    ", \"name\": " + "\"" + name + "\"" +
+                    ", \"quantity\": " + "\"" + quantity + "\"" +
+                    ", \"description\": " + "\"" + description + "\""  +
+                    '}';
+        }
     }
 }
