+++ src/main/java/managers/StaffManager.java
@@ -1,4 +1,4 @@
-package p;
+package managers;
 
 import org.json.*;
