+++ src/main/java/tests/managers/OrderManagerTest.java
@@ -1,0 +1,18 @@
+package tests.managers;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class OrderManagerTest {
+
+    @org.junit.jupiter.api.Test
+    void addJsonObj() {
+    }
+
+    @org.junit.jupiter.api.Test
+    void removeJsonObj() {
+    }
+
+    @org.junit.jupiter.api.Test
+    void searchJsonObj() {
+    }
+}
